Splitting utils into different modules : - Metadata and Metadatum have their own common.metadata namespace - Reference, Citation and URN have their own common.reference namespace - NS constant and parse (renamed xmlparser) from resources.xml moved to own common.utils namespace
