+++ MyCapytain/common/reference.py
@@ -1,5 +1,7 @@
 from __future__ import unicode_literals
-from collections import defaultdict, OrderedDict
+
+
+from collections import defaultdict
 from past.builtins import basestring
 from builtins import range, object
 import re
@@ -7,9 +9,7 @@ import re
 
 REFSDECL_SPLITTER = re.compile("/+[a-zA-Z0-9:\[\]@=\\\{\$'\"\.]+")
 REFSDECL_REPLACER = re.compile("\$[0-9]+")
-"""
-    The ```utils``` modules
-"""
+SUBREFERENCE = re.compile("(\w*)\[{0,1}([0-9]*)\]{0,1}", re.UNICODE)
 
 class Reference(object):
 
@@ -75,8 +75,7 @@ class Reference(object):
         :rtype: List.<Tuple>
         :returns: List where first element is a tuple representing different components
         """
-        r = re.compile("(\w*)\[{0,1}([0-9]*)\]{0,1}", re.UNICODE)
-        return r.findall(subreference)[0]
+        return SUBREFERENCE.findall(subreference)[0]
 
     def __parse(self, reference):
         """ Parse references informations
@@ -303,143 +302,6 @@ class URN(object):
         else:
             raise ValueError("URN is empty")
         return parsed
-
-class Metadatum(object):
-
-    """ Metadatum object represent a single field of metadata """
-
-    def __init__(self, name, children=None):
-        """ Initiate a Metadatum object
-        :param name: Name of the field
-        :type name: basestring
-        :param children: List.tuple
-        :type children: dict
-        """
-        self.name = name
-        self.children = OrderedDict()
-        self.default = None
-
-        if children is not None and isinstance(children, list):
-            for tup in children:
-                self[tup[0]] = tup[1]
-
-    def __getitem__(self, key):
-        """ Add an iterable access method
-        :param key:
-        :type key:
-        :returns: An element of children whose index is key
-        """
-        if isinstance(key, int):
-            items = list(self.children.keys())
-            if key + 1 > len(items):
-                raise KeyError()
-            else:
-                key = items[key]
-        elif isinstance(key, tuple):
-            return tuple([self[k] for k in key])
-
-        if key not in self.children:
-            if self.default is None:
-                raise KeyError()
-            else:
-                return self.children[self.default]
-        else:
-            return self.children[key]
-
-    def __setitem__(self, key, value):
-        """ Add an iterable access method
-        :param key:
-        :type key:
-        :param value:
-        :type value:
-        :returns: An element of children whose index is key
-        """
-        if isinstance(key, tuple):
-            if not isinstance(value, (tuple, list)):
-                value = [value]*len(key)
-            if len(value) < len(key):
-                raise ValueError("Less values than keys detected")
-            for i in range(0, len(key)):
-                self[key[i]] = value[i]
-        elif not isinstance(key, basestring):
-            raise TypeError("Only basestring or tuple instances are accepted as key")
-        else:
-            self.children[key] = value
-            if self.default is None:
-                self.default = key
-
-    def setDefault(self, key):
-        if key not in self.children:
-            raise ValueError("Can not set a default to an unknown key")
-        else:
-            self.default = key
-        return self.default
-
-    def __iter__(self):
-        i = 0
-        for key in self.children:
-            yield (key, self.children[key])
-            i+=1
-
-class Metadata(object):
-    def __init__(self, keys=None):
-        self.metadata = defaultdict(Metadatum)
-        self.__keys = []
-        if keys is not None:
-            for key in keys:
-                self[key] = Metadatum(name=key)
-
-    def __getitem__(self, key):
-        """ Add an iterable access method
-        :param key:
-        :type key:
-        :returns: An element of children whose index is key
-        """
-        if isinstance(key, int):
-            if key + 1 > len(self.__keys):
-                raise KeyError()
-            else:
-                key = self.__keys[key]
-        elif isinstance(key, tuple):
-            return tuple([self[k] for k in key])
-
-        if key not in self.metadata:
-            raise KeyError()
-        else:
-            return self.metadata[key]
-
-    def __setitem__(self, key, value):
-        """ Add an iterable access method
-        :param key:
-        :type key:
-        :param value:
-        :type value:
-        :returns: An element of children whose index is key
-        """
-        if isinstance(key, tuple):
-            if len(value) < len(key):
-                raise ValueError("Less values than keys detected")
-            for i in range(0, len(key)):
-                self[key[i]] = value[i]
-        elif not isinstance(key, basestring):
-            raise TypeError("Only basestring or tuple instances are accepted as key")
-        else:
-            if not isinstance(value, Metadatum) and isinstance(value, list):
-                self.metadata[key] = Metadatum(key, value)
-            elif isinstance(value, Metadatum):
-                self.metadata[key] = value
-
-            if key in self.metadata and key not in self.__keys:
-                self.__keys.append(key)
-
-    def __iter__(self):
-        i = 0
-        for key in self.__keys:
-            yield (key, self.metadata[key])
-            i+=1
-
-    def __len__(self):
-        return len([k for k in self.__keys if isinstance(self.metadata[k], Metadatum)])
 
 
 class Citation(object):
