--- MyCapytain/parser.py
@@ -1,3 +1,0 @@
-"""
-  Reply parser for CTS API's answers
-"""
